Update globalization-invariant-mode.md
On this page under ["Revert back to NLS behaviors" ](https://docs.microsoft.com/en-us/dotnet/standard/base-types/string-comparison-net-5-plus) it sais that the ICU setting has to be set for the whole application. Therefor adding the setting for each project (as suggested at point 1.) wouldn't be handy, as they all have to have the same setting anyway.
This would be in the Directory.Build.props file. 

	<Project>
		<PropertyGroup>
			<InvariantGlobalization>false</InvariantGlobalization>
		</PropertyGroup>
	</Project>